requirements.txt
asgiref==3.7.2
chardet==5.2.0
Django==4.0
django-cors-headers==4.3.1
django-el-pagination==4.0.0
djangorestframework==3.14.0
numpy==1.26.4
pandas==2.2.0
pillow==10.2.0
python-dateutil==2.8.2
pytz==2024.1
reportlab==4.0.9
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
